fix: 검색 결과 페이지에서 새로고침했을 때 헤더 없어지는 문제 해결하기
