+++ src/pages/searchResult/style.css.ts
@@ -13,16 +13,6 @@ export const wrapper = style([
 
 export const fetching = style({ color: variants.color.primary });
 
-export const searchInfo = style({
-  marginBottom: '4.2rem',
-});
-
-export const resultStats = style({
-  fontSize: variants.fontSize.medium,
-  color: variants.color.font.secondary,
-  marginBottom: '1.4rem',
-});
-
 export const noResult = style([
   utils.textAlignCenter,
   {
